+++ test/system/sor_test.dart
@@ -47,6 +47,28 @@ void main() {
       expect(solutions[2], MoreOrLessEquals(-2, precision: 1.0e-2));
     });
 
+    test("Making sure that the string conversion works properly.", () {
+      final solver = SORSolver(
+        equations: const [
+          [3, -1, 1],
+          [-1, 3, -1],
+          [1, -1, 3]
+        ],
+        constants: [-1, 7, -7],
+        w: 1.25,
+      );
+
+      final toString = "[3.0, -1.0, 1.0]\n"
+          "[-1.0, 3.0, -1.0]\n"
+          "[1.0, -1.0, 3.0]";
+      final toStringAugmented = "[3.0, -1.0, 1.0 | -1.0]\n"
+          "[-1.0, 3.0, -1.0 | 7.0]\n"
+          "[1.0, -1.0, 3.0 | -7.0]";
+
+      expect(solver.toString(), equals(toString));
+      expect(solver.toStringAugmented(), equals(toStringAugmented));
+    });
+
     test(
         "Making sure that the matrix is squared AND the dimension of the "
         "known values vector also matches the size of the matrix.", () {